requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
tensorflow==2.10.1
keras==2.10.0
pillow
numpy